refactor: chess package 안으로 Pawn, Board 파일 이동
