+++ src/main/java/softeer2nd/chess/Pawn.java
@@ -1,4 +1,4 @@
-package softeer2nd;
+package softeer2nd.chess;
 
 public class Pawn {
     public static final String WHITE = "white";
