+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -1,4 +1,4 @@
-package softeer2nd;
+package softeer2nd.chess;
 
 import org.junit.jupiter.api.Test;
